+++ react-app/src/components/User.js
@@ -51,7 +51,8 @@ function User() {
   }, [userId]);
 
   const deleteBtn = async rec_id => {
-    await dispatch(deleteRecordThunk(rec_id));
+    if (window.confirm("Are you sure you want to delete this record?"))
+      await dispatch(deleteRecordThunk(rec_id));
   };
 
   if (!user) {
